Move test_sparse_class_distribution to utils/multiclass from utils/sparsefuncs
